Изменил типы данных в свойствах всех сущностей, Разделил сущность User на сущность User и UserProfile

+++ SQLServerCourse.Domain/EntitiesForPracticalTasks/HallRow.cs
@@ -9,9 +9,9 @@ namespace SQLServerCourse.Domain.Entitys_for_lesson
 {
     public class HallRow
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int HallId { get; set; }
+        public long HallId { get; set; }
 
         public short Number { get; set; }
