+++ SQLServerCourse.Domain/Entity/Question.cs
@@ -9,9 +9,9 @@ namespace SQLServerCourse.Domain.Entity
 {
     public class Question
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
 
-        public int LessonId { get; set; }
+        public byte LessonId { get; set; }
 
         public Lesson Lesson { get; set; }
